refactor(search): split station_card_parts into 3 sibling part files (Refs #563)
The 360-LOC station_card_parts.dart exceeded the 300-LOC presentation
guideline. Extract its three logical sections into sibling part files of
station_card.dart so each piece stays reviewable: status (status dot +
station details), price column (selected-fuel price + favorite + cheapest
badge), and the alt-fuel price row used in all-fuels mode.

Public StationCard API is unchanged; the existing widget test suite passes
without modification.
